feature: add UserDetailModal and CreateUserModal
