+++ components/TreatmentMenu.tsx
@@ -26,7 +26,7 @@ export default function TreatmentMenu() {
                 setTreatments(data);
             } catch (error) {
                 setError(
-                    'No se pudieron cargar los tratamientos. Inténtalo de nuevo más tarde.',
+                    'Error al cargar los tratamientos. Inténtalo de nuevo más tarde.',
                 );
             }
         };
@@ -43,7 +43,7 @@ export default function TreatmentMenu() {
 
     return (
         <>
-            <div>{error && <p style={{ color: 'red' }}>{error}</p>}</div>
+            <div >{error && <p style={{ color: 'red',padding:'10px' }}>{error}</p>}</div>
             <SelectRoot
                 onValueChange={(e) => handleTreatmentSelect(e.value.toString())}
                 size='lg'
